pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "quickstart-pytorch"
version = "1.0.0"
description = "Federated Learning with PyTorch and Flower (Quickstart Example)"
license = "Apache-2.0"
dependencies = [
    "flwr[simulation]>=1.24.0",
    "flwr-datasets[vision]>=0.5.0",
    "torch==2.8.0",
    "torchvision==0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "flwrlabs"

[tool.flwr.app.components]
serverapp = "pytorchexample.server_app:app"
clientapp = "pytorchexample.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 5
fraction-evaluate = 1.0
local-epochs = 5
learning-rate = 0.01
batch-size = 64
malicious-nodes = 0 # 悪意のあるクライアントの数

[tool.flwr.federations]
default = "local-simulation-gpu"

[tool.flwr.federations.local-simulation]
options.num-supernodes = 10

[tool.flwr.federations.local-simulation-gpu]
options.num-supernodes = 10
options.backend.client-resources.num-cpus = 2 # each ClientApp assumes to use 2CPUs
options.backend.client-resources.num-gpus = 0.2 # at most 5 ClientApp will run in a given GPU

[tool.flwr.federations.supergrid-deployment]
address = "supergrid.flower.ai"
enable-account-auth = true